feat: moved constants
