+++ src/components/pages/EmployeesSkillsPage/constants/skillsMastery.ts
@@ -1,6 +1,5 @@
 export enum SkillsMastery {
   Novice = 'novice',
-  Average = 'average',
   Advanced = 'advanced',
   Expert = 'expert',
 }
