+++ src/hooks/useEmployeeSkillsFormData.ts
@@ -3,7 +3,7 @@ import { SKILLS } from '../graphql/queries/skills';
 import { USER } from '../graphql/queries/user';
 import { ISkillsReturn } from '../graphql/types/results/skills';
 import { IUserAllResult } from '../graphql/types/results/user';
-import { SkillsMastery } from './../constants/skillsMastery';
+import { SkillsMastery } from '../components/pages/EmployeesSkillsPage/constants/skillsMastery';
 
 export const useEmployeeSkillsFormData = (id: string) => {
   const { loading: loadingUser, error: errorUser, data: userData } = useQuery<IUserAllResult>(
